+++ src/azure-cli/HISTORY.rst
@@ -20,6 +20,7 @@ Release History
 **Compute**
 
 * Upgrade API version of compute, disks, snapshots to 2019-07-01
+* vmss create: Improvement for --orchestration-mode
 
 **IOT**